botz/requirements.txt
python-telegram-bot>=20.0
requests
gtts
SpeechRecognition
ffmpeg-python
gunicorn
aiohttp 